Partially port basic_subscriber sample.
This commit contains an example that acts much like the basic_subscriber
example in the main Aeron project; however, rather than consuming a
"hello world" string, it consumes a message with three 32-bit integer
values. This deviation is because the `Unsafe_buffer` has only limited
functionality, at the time of writing.

It may be used in conjunction with the basic_publisher example.