Avoid int32 when (native) int will do.
In OCaml, it is more efficient to use `int` than `int32` on 64-bit
machines where you only need 63-bits, as `int32` is represented as a
boxed value, i.e., a pointer to a block on the heap where the block has
a size, a tag, and the value.

In this commit, I've replaced `int32` with `int` for:

1. stream id,
2. buffer slice offset,
3. buffer slice length,
4. resulting publication offer position, and
5. number of fragments consumed during a poll.

All of these replacements should be safe except for (4) when a
stream is _very_ long. I'm happy to accept this limitation.

Note, stream id should be in the `int32` range, we should consider
adding some defensive coding here.